requirements.txt
# Django and Web Framework
Django==4.2.8
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9

# Image Processing
Pillow==10.1.0

# Multilingual Support
django-modeltranslation==0.18.11
django-parler==2.3

# Forms and Validation
django-crispy-forms==2.1
crispy-tailwind==0.5.0

# Payment Integration
stripe==7.7.0

# Admin Enhancements
django-jazzmin

# File Cleanup
django-cleanup==8.0.0

# Rich Text Editor
django-tinymce==3.6.0

# SEO and Meta Tags
django-meta==2.3.0

# Utilities
python-slugify==8.0.1
python-dateutil==2.8.2

# Development Tools
django-debug-toolbar==4.2.0
ipython==8.18.1

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# Email
django-anymail==10.2

# Security
django-cors-headers==4.3.1

# Test Data Generation
Faker==20.1.0

